starpal/starpal_project/starpal_project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
openai==1.6.1
python-dotenv
mem0ai